Rewrite cgoase function to library functions in libase/term
