+++ libase/term/doc.go
@@ -1,0 +1,2 @@
+// aseterm contains common code for cgoase and goase.
+package term
